Technically works
the room 2 monster generation is bugged pretty hard, partially because the random monster choice isn't carried properly and the text won't match the monster. There were also many things that I had added/fixed in other versions but unfortunately when I tried to fix room 2 everything broke. This is the "original" copy of the game.